+++ R/read_hpa_spatial_data.R
@@ -14,7 +14,7 @@ read_hpa_spatial_data <- function(name, export_dir = tempdir()) {
   # adapted from strayr::read_absmap()
   assertthat::assert_that(assertthat::is.writeable(export_dir))
 
-  base_url <- "https://github.com/healthpolicyanalysis/hpa.spatial.data/raw/3e657efb8d5a8a2a4b97ba67871923a02b14314d/data/"
+  base_url <- "https://github.com/healthpolicyanalysis/hpa.spatial.data/raw/fcdd8ccc7f87ba3a48f527180c84db4ea6a21ad4/data/"
   out_path <- file.path(export_dir, paste0(name, ".rda"))
 
   if (!file.exists(out_path)) {
@@ -69,7 +69,8 @@ get_data_file_list <- function() {
     "phn",
     "mb21_poly",
     "mb21_pop",
-    "mmm19"
+    "mmm19",
+    "mmm23"
   )
 }